The documentation and the configuration of a Replication Domain allow a maximum value of 65535 for the server-id property. Nevertheless, the server-id in the ReplicationDomain implementation is managed as a short allowing a maximum value of 32767.
With this change the code now uses an int to store the server-id.
The maximum is still limited to 65535 and this is enforced by the management framework.

This change should not impact compatibility as the messages exchanged by the servers
are not impacted.

This change also add unit tests allowing to test the compatibility of the V4 protocol
with V3 protocol.

+++ opendj-sdk/opends/src/server/org/opends/server/replication/common/MultiDomainServerState.java
@@ -86,7 +86,7 @@ public class MultiDomainServerState implements Iterable<String>
 
     synchronized(this)
     {
-      Short serverId =  changeNumber.getServerId();
+      int serverId =  changeNumber.getServerId();
       ServerState oldServerState = list.get(serviceId);
       if (oldServerState == null)
         oldServerState = new ServerState();
